Global refactoring, add more dependencies
