+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,8 @@
         android:theme="@style/Theme.DanceClub"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.DanceClub">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
