+++ app/src/main/java/com/example/danceclub/data/local/dao/SectionDao.kt
@@ -1,9 +1,10 @@
-package com.example.danceclub.section
+package com.example.danceclub.data.local.dao
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import com.example.danceclub.data.model.Section
 
 @Dao
 interface SectionDao { // тут не все, что умеет делать секция
