+++ app/src/main/java/com/example/danceclub/data/model/Section.kt
@@ -1,4 +1,4 @@
-package com.example.danceclub.section
+package com.example.danceclub.data.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -7,11 +7,11 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "sections")
 data class Section(
     @PrimaryKey(autoGenerate = true)
-    var id:Int,
+    var id: Int,
     @ColumnInfo(name = "info")
-    var info:String,
+    var info: String,
     @ColumnInfo(name = "name")
     var name: String,
     @ColumnInfo(name = "isFree")
-    var isFree:Boolean
+    var isFree: Boolean
 )
