+++ app/src/main/java/com/example/danceclub/ui/MainActivity.kt
@@ -1,0 +1,24 @@
+package com.example.danceclub.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.material3.Surface
+import com.example.danceclub.ui.theme.DanceClubTheme
+
+class MainActivity : ComponentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        setContent {
+            DanceClubTheme {
+                Surface {
+                    DanceAppNavHost()
+                }
+            }
+        }
+    }
+}
