+++ app/src/main/java/com/example/danceclub/ui/navigation/NavigationDirections.kt
@@ -1,0 +1,18 @@
+package com.example.danceclub.ui.navigation
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+object GreetingDestination
+
+@Serializable
+object RegistrationDestination
+
+@Serializable
+object SignInDestination
+
+@Serializable
+object ProfileDestination
+
+@Serializable
+object TrainingsDestination
