+++ app/src/main/java/com/example/danceclub/ui/screens/auth/sing_in/SignInViewModel.kt
@@ -1,20 +1,21 @@
-package com.example.danceclub
+package com.example.danceclub.ui.screens.auth.sing_in
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
-import com.example.danceclub.account.Account
-import com.example.danceclub.account.AccountsDao
-import com.example.danceclub.section.SectionDao
+import com.example.danceclub.data.model.Account
+import com.example.danceclub.data.local.dao.AccountsDao
+import com.example.danceclub.data.local.AppDatabase
+import com.example.danceclub.data.local.dao.SectionDao
 
-class AuthorizationViewModel(application: Application) : AndroidViewModel(application) {
+class SignInViewModel(application: Application) : AndroidViewModel(application) {
     private val accountDao: AccountsDao = AppDatabase.getInstance(application).accountsDao()
     private val sectionsDao: SectionDao = AppDatabase.getInstance(application).sectionsDao()
     private val appDatabase: AppDatabase = AppDatabase.getInstance(application)
 
-
     fun findAccount(username: String): Account? {
         return accountDao.searchAccount(username)
     }
+
     /*fun getSections(): LiveData<List<Section>> {
         return appDatabase.sectionsDao().getSections()
     }
@@ -22,11 +23,11 @@ class AuthorizationViewModel(application: Application) : AndroidViewModel(applic
     fun getAccounts(): LiveData<List<Account>> {
         return appDatabase.accountsDao().getAccounts()
     }
+
     fun saveSection(section: Section) {
         sectionsDao.add(section)
     }
 
      */
-
 
 }
