+++ app/src/main/java/com/example/danceclub/ui/screens/trainings/TrainingsScreen.kt
@@ -1,0 +1,10 @@
+package com.example.danceclub.ui.screens.trainings
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun TrainingsScreen(
+    onNavigateUpToProfile: () -> Unit,
+) {
+    // TODO
+}
