+++ build.gradle.kts
@@ -2,4 +2,6 @@
 plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.jetbrains.kotlin.android) apply false
+    alias(libs.plugins.ksp) apply false
+    alias(libs.plugins.jetbrains.kotlin.serialization) apply false
 }
